✨ ajustes na estrutura, camada controller de source protegida
